+++ src/main/scala/org/podval/tools/scalajsplugin/jvm/JvmTestEnvironment.scala
@@ -1,5 +1,6 @@
 package org.podval.tools.scalajsplugin.jvm
 
+import org.podval.tools.build.ScalaBackendKind
 import org.podval.tools.test.environment.{SourceMapper, TestEnvironment}
 import org.podval.tools.test.framework.{FrameworkDescriptor, FrameworkProvider}
 import sbt.testing.Framework
@@ -11,5 +12,6 @@ final class JvmTestEnvironment extends TestEnvironment:
 
   protected def expandClassPath: Boolean = true
   
-  override protected def loadFrameworks: List[Framework] =
-    FrameworkDescriptor.jvmSupported.flatMap(FrameworkProvider(_).frameworkOpt)
+  override protected def loadFrameworks: List[Framework] = FrameworkDescriptor
+    .forBackend(ScalaBackendKind.JVM)
+    .flatMap(FrameworkProvider(_).frameworkOpt)