+++ src/main/scala/org/podval/tools/scalajsplugin/scalajs/ScalaJSTestTask.scala
@@ -1,5 +1,6 @@
 package org.podval.tools.scalajsplugin.scalajs
 
+import org.podval.tools.build.ScalaBackendKind
 import org.podval.tools.test.task.TestTask
 
 abstract class ScalaJSTestTask extends TestTask with ScalaJSRunTask:
@@ -7,6 +8,6 @@ abstract class ScalaJSTestTask extends TestTask with ScalaJSRunTask:
 
   final override protected def linkTaskClass: Class[ScalaJSLinkTestTask] = classOf[ScalaJSLinkTestTask]
 
-  final override protected def isScalaJS: Boolean = true
+  final override protected def backendKind: ScalaBackendKind = ScalaBackendKind.JS
 
   final override protected def createTestEnvironment: ScalaJSTestEnvironment = ScalaJSTestEnvironment(scalaJSRunCommon)